pyproject.toml
[project]
name = "fast-model-loading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sagemaker",
    "boto3",
    "zmq",
    "python-dotenv",
    "huggingface-hub>=0.34.4",
]
